requirements.txt
pyyaml
requests
Pillow
esptool
